# docs/TalkRoomsOpenApi.yaml
openapi: 3.0.1
info:
  title: TalkRooms API
  description: 'System to manage message channels and send messages'
  termsOfService: https://swagger.io/terms/
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /user/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Users
      summary: Get details of a user
      description: Returns a single user
      parameters:
        - name: id
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Username missing or invalid
        '401':
          description: Unauthorized

        '404':
          description: User not found
        '500':
          description: Internal server error
...
  /channels:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Create a channel
      requestBody:
        description: Details of channel
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelInputModel'
      responses:
        '201':
          description: Channel registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel name
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
  /channels/{channelId}/add/{userId}/{role}:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Add a user to a channel
      parameters:
        - name: channelId
          in: path
          description: channel identifier
          required: true
          schema:
            type: integer
            example: 1
        - name: userId
          in: path
          description: user identifier
          required: true
          schema:
            type: integer
            example: 1
        - name: role
          in: path
          description: user role
          required: true
          schema:
            type: string
            description: Role of the user in the channel
            enum: [ADMIN, READ_ONLY, READ_WRITE]
            example: READ_ONLY
      responses:
        '200':
          description: User added to channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel
        '401':
          description: Unauthorized
        '404':
          description: Channel not found
        '500':
          description: Internal server error
  /channels/{id}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Get details of a channel
      description: Returns a single channel
      parameters:
        - name: id
          in: path
          description: Channel identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Channel found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel
        '401':
          description: Unauthorized
        '404':
          description: Channel not found
        '500':
          description: Internal server error
  /channels/search/{name}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Search channel by name
      description: Returns a list of channel
      parameters:
        - name: name
          in: path
          description: Channel name
          required: true
          schema:
            type: string
            example: DAW
        - name: limit
          in: query
          description: number of channels to return
          required: false
          schema:
            type: integer
            example: 20
        - name: skip
          in: query
          description: number of channels to skip
          required: false
          schema:
            type: integer
            example: 0
      responses:
        '200':
          description: Channel found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel
        '401':
          description: Unauthorized
        '404':
          description: Channel not found
        '500':
          description: Internal server error
  /channels/{channelId}/members:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Get channels by user id
      description: Returns a list of channels
      parameters:
        - name: channelId
          in: path
          description: channel identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Members found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '400':
          description: Invalid user
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
  /channels/user/{userId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Get channels by user id
      description: Returns a list of channels
      parameters:
        - name: userId
          in: path
          description: user identifier
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Channels found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelList'
        '400':
          description: Invalid user
        '401':
          description: Unauthorized
        '404':
          description: User not found
        '500':
          description: Internal server error
  /channels/{channelId}/{name}:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Update channel name
      description: Create a new channel given a name and a visibility
      parameters:
        - name: channelId
          in: path
          description: Channel identifier
          required: true
          schema:
            type: integer
            example: 1
        - name: name
          in: path
          description: Channel new name
          required: true
          schema:
            type: string
            example: PDM
      responses:
        '200':
          description: Channel created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel data
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /channels/{channelId}/leave/{userId}:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Channels
      summary: Remove user from channel
      description: Remove user from channel
      parameters:
        - name: channelId
          in: path
          description: Channel identifier
          required: true
          schema:
            type: integer
            example: 1
        - name: userId
          in: path
          description: User identifier
          required: true
          schema:
            type: integer
            example: 23
      responses:
        '200':
          description: User removed from channel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelOutputModel'
        '400':
          description: Invalid channel data
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserIdentifiers:
      type: object
      description: User details
      required:
        - id
        - username
      properties:
        id:
          type: integer
          description: user id
          example: 1
        username:
          type: string
          description: username
          example: Bob123
    CreateChannelInputModel:
      type: object
      description: Data to create channel
      required:
        - name
        - visibility
      properties:
        name:
          type: string
          description: channel name
          example: DAW
        visibility:
          type: string
          description: channel visibility
          enum: [PUBLIC, PRIVATE]
          example: PUBLIC
    ChannelOutputModel:
      type: object
      description: Data that represents a channel
      required:
        - id
        - name
        - creator
        - visibility
      properties:
        id:
          type: integer
          description: channel identifier
          example: 1
        creator:
          $ref: '#/components/schemas/UserIdentifiers'
        name:
          type: string
          description: channel name
          example: DAW
        visibility:
          type: string
          description: channel visibility
          enum: [PUBLIC, PRIVATE]
          example: PUBLIC
    ChannelList:
      type: object
      properties:
        nChannels:
          type: integer
          description: number of results
          example: 20
        channels:
          type: array
          items:
            $ref: '#/components/schemas/ChannelOutputModel'
...
    UserList:
      type: object
      properties:
        nUsers:
          type: integer
          description: number of results
          example: 20
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserIdentifiers'
...
    User:
      type: object
      description: User details
      required:
        - id
        - username
        - email
      properties:
        id:
          type: integer
          description: user identifier
        username:
          type: string
          description: user username
          example: Bob
        email:
          type: string
          description: user email
          example: bob@example.pt
...